dp_mixgauss sampling jaxified, fig 30.4
